+++ HISTORY.rst
@@ -5,8 +5,10 @@ Changelog
 ******************
 
 - Prompts variables can have ANSI colors and styles.
-- Prompt variable``{{full_cwd}}`` renamed to ``{{cwd}}.
+- Prompt variable``{{full_cwd}}`` renamed to ``{{cwd}}``.
 - Prompt variable ``{cwd}`` renamed to ``{dir}``.
+- Short option for `--speed` is now ``-s``.
+- Short option for ``--shell`` is now ``-S``.
 - ``{hostname}`` can be included in prompt.
 - ``run`` and ``magictype`` receive prompt_template instead of a prompt function.
 - Remove unnecessary ``PromptState`` class.